+++ singlem/lyrebird.py
@@ -150,6 +150,8 @@ def main():
             exclude_off_target_hits = args.exclude_off_target_hits,
             min_taxon_coverage = get_min_taxon_coverage(args),
             max_species_divergence = args.max_species_divergence,
+            repair_frameshifts = not args.no_repair_frameshifts,
+            max_frameshift_repair_divergence = args.max_frameshift_repair_divergence,
         )
 
     elif args.subparser_name=='renew':